lowering: move expr builder methods -> expr.rs
